Added `remap_filters` method to `LightCurve` and use in SNANA
Adding a static method `remap_filters` which changes the names of the
filters in the `.lightCurve` attribute. This requires a dictionary as
intput mapping the old values of the filters to the new values of the
filter.

The dictionary can be provided as an input to the `LightCurve` constructor
along with a boolean `ignore_case` which ignores the case for the
filter.

In `SNANASims` a dictionary attribute is added as a constant, which maps
the SNANA filter values of 'ugrizY' to 'lsst_u', etc. Then the
`get_SNANAPhotometry` method automatically provides strings that use the
LSST filters registered in `.filters.py`